Updated to alpha 28.4

+++ src/App/Osu/OsuBeatmapDifficulty.h
@@ -119,7 +119,7 @@ public:
 	UString backgroundImageName;
 
 	unsigned long previewTime;
-	unsigned long lastModificationTime;
+	long long lastModificationTime;
 
 	// objects
 	std::vector<HITCIRCLE> hitcircles;
